*More code cleanup.  Moving variable declarations before program code in each code block, because Splint doesn't support the C99 standard of allowing variables to be declared after the first line of code in a function. *Removed the commented out eof_editor_logic_feedback() function, as it is fully deprecated.

+++ trunk/src/main.h
@@ -246,7 +246,7 @@ extern void      * eof_music_data;
 extern int         eof_silence_loaded;	//Tracks whether "second_of_silence.ogg" was loaded from EOF's program folder instead of user-specified chart audio
 extern int         eof_music_data_size;
 extern unsigned long eof_chart_length;
-extern int         eof_music_length;
+extern unsigned long eof_music_length;
 extern int         eof_music_pos;
 extern int         eof_music_actual_pos;
 extern int         eof_music_rewind_pos;
